refactor(core): move to modules
