+++ src/app/modules/xm-core/index.ts
