+++ src/app/modules/xm-core/src/config/xm-private-ui-config.service.ts
