+++ src/app/modules/xm-core/src/config/xm-public-ui-config.service.spec.ts
