+++ src/app/modules/xm-core/src/config/xm-ui-config-model.ts
