+++ src/app/modules/xm-core/src/xm-core-config.ts
