+++ src/app/modules/xm-core/src/xm-core.module.ts
@@ -3,9 +3,9 @@ import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { defaults } from 'lodash';
 import { JhiEventManager } from 'ng-jhipster';
-import { ErrorHandlerInterceptor } from '../../../modules/xm-core/src/errorhandler.interceptor';
-import { XmPermissionService } from '../../privilege/xm-permission.service';
-import { XmEventManager } from '../index';
+import { ErrorHandlerInterceptor } from './errorhandler.interceptor';
+import { XmPermissionService } from '../../../shared/privilege/xm-permission.service';
+import { XmEventManagerService as XmEventManager } from './xm-event-manager.service';
 import { XmUiConfigService } from './config/xm-ui-config.service';
 import { XM_CORE_EXTERNAL_CONFIG, XmCoreConfig } from './xm-core-config';
 import { XmEventManagerService } from './xm-event-manager.service';
