+++ src/app/modules/xm-core/src/xm-event-manager.service.spec.ts
