+++ src/app/modules/xm-core/src/xm-event-manager.service.ts
