+++ src/app/modules/xm-core/src/xm-session.service.ts
