+++ tsconfig.json
@@ -21,7 +21,7 @@
     "paths": {
       "@xm-ngx/styles": ["packages/styles"],
       "@xm-ngx/components/*": ["packages/components/src/*"],
-      "@xm-ngx/core": ["src/app/shared/core"],
+      "@xm-ngx/core": ["src/app/modules/xm-core"],
       "@xm-ngx/core/auth": ["src/app/modules/xm-core-auth"],
       "@xm-ngx/ext/*": ["src/app/ext/*"],
       "@xm-ngx/core/environment": ["src/environments"],
